+++ client/src/router/NeedUserLogin.tsx
@@ -10,7 +10,7 @@ import SettingsCsvPage from '../pages/settings-csv-page/SettingsCsvPage';
 import SettingsSocialPage from '../pages/settings-social-page/SettingsSocialPage';
 import StatisticsPage from '../pages/statistics-page/StatisticsPage';
 import TransactionPage from '../pages/transaction-page/TransactionPage';
-import LoginGuard from './LoginGuard';
+import LoginGuard from '../hoc/LoginGuard';
 
 const NeedUserLogin: React.FC = () => {
   useGetAuthority();
